+++ src/hellotest/Hello.java
@@ -11,10 +11,10 @@ import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.MediaType;
 import javax.ws.rs.core.Response;
 
-import us.kbase.auth2.service.exceptions.AuthException;
-import us.kbase.auth2.service.exceptions.AuthenticationException;
-import us.kbase.auth2.service.exceptions.AuthError;
-import us.kbase.auth2.service.exceptions.UnauthorizedException;
+import us.kbase.auth2.lib.exceptions.AuthError;
+import us.kbase.auth2.lib.exceptions.AuthException;
+import us.kbase.auth2.lib.exceptions.AuthenticationException;
+import us.kbase.auth2.lib.exceptions.UnauthorizedException;
 
 
 // Plain old Java Object it does not extend as class or implements
