+++ src/hellotest/HelloApp.java
@@ -11,7 +11,7 @@ import org.slf4j.LoggerFactory;
 import ch.qos.logback.classic.Level;
 import ch.qos.logback.classic.Logger;
 import us.kbase.auth2.service.LoggingFilter;
-import us.kbase.auth2.service.exceptions.handler.ExceptionHandler;
+import us.kbase.auth2.service.exceptions.ExceptionHandler;
 import us.kbase.auth2.service.template.TemplateProcessor;
 import us.kbase.auth2.service.template.mustache.MustacheProcessor;
