+++ src/us/kbase/auth2/lib/exceptions/AuthException.java
@@ -1,4 +1,4 @@
-package us.kbase.auth2.service.exceptions;
+package us.kbase.auth2.lib.exceptions;
 
 /** Base class of all authorization service exceptions.
  * @author gaprice@lbl.gov
