+++ src/us/kbase/auth2/lib/storage/AuthStorage.java
@@ -2,12 +2,12 @@ package us.kbase.auth2.lib.storage;
 
 import java.util.List;
 
-import us.kbase.auth2.lib.HashedToken;
 import us.kbase.auth2.lib.LocalUser;
+import us.kbase.auth2.lib.exceptions.AuthException;
+import us.kbase.auth2.lib.exceptions.AuthenticationException;
 import us.kbase.auth2.lib.storage.exceptions.AuthStorageException;
+import us.kbase.auth2.lib.token.HashedToken;
 import us.kbase.auth2.lib.token.IncomingHashedToken;
-import us.kbase.auth2.service.exceptions.AuthException;
-import us.kbase.auth2.service.exceptions.AuthenticationException;
 
 public interface AuthStorage {
