+++ src/us/kbase/auth2/lib/token/AuthToken.java
@@ -1,4 +1,4 @@
-package us.kbase.auth2.lib;
+package us.kbase.auth2.lib.token;
 
 import static us.kbase.auth2.lib.Utils.checkString;
