+++ src/us/kbase/auth2/lib/token/HashedToken.java
@@ -1,4 +1,4 @@
-package us.kbase.auth2.lib;
+package us.kbase.auth2.lib.token;
 
 import java.util.Date;
 import java.util.UUID;
