+++ src/us/kbase/auth2/lib/token/IncomingToken.java
@@ -1,7 +1,7 @@
 package us.kbase.auth2.lib.token;
 
 import static us.kbase.auth2.lib.Utils.checkString;
-import static us.kbase.auth2.lib.AuthToken.hash;
+import static us.kbase.auth2.lib.token.AuthToken.hash;
 
 public class IncomingToken {
